Unified TestLib and TestApp in favor of TestHelper
