+++ packages/cli/src/index.js
@@ -2,7 +2,7 @@
 const version = 'v' + require('../package.json').version
 
 // model objects
-import { TestApp, LibApp }from './models/TestApp'
+import TestHelper from './models/TestHelper'
 import ControllerFor from './models/local/ControllerFor'
 import LocalAppController from './models/local/LocalAppController'
 import LocalLibController from './models/local/LocalLibController'
@@ -11,8 +11,7 @@ import NetworkLibController from './models/network/NetworkLibController'
 
 export {
   version,
-  TestApp,
-  LibApp,
+  TestHelper,
   ControllerFor,
   LocalAppController,
   LocalLibController,
