+++ packages/cli/test/mocks/packages/package-with-contracts-and-multiple-stdlibs.json
@@ -1,6 +1,6 @@
 {
   "name": "Herbs",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "contracts": {
     "Impl": "ImplV1"
   },
